LibGfx: Enable in Windows CI
When building, I was getting a link error as we were trying to link
with the 'm.lib' import lib. But having to explicitly link in math
symbols is for Unix only. It turns out that the libjxl dependency
either itself or one of its dependencies is transitively linking
with libm (something somewhere is doing PUBLIC/INTERFACE m).

From looking at the libjxl repo it feels very unix-y (their build
instructions just point to a shell script), so I guess JXL is only
built on windows via MingW64 or something? In that case then maybe
it'd be the vcpkg port authors exposed added windows support that
would need to fix it, or we just indefinitely keep this hack