CMake: Add Windows executable helper function
The function currently has 2 purposes: (1) To copy dependent dlls
for executables to output binary directory. This ensures that these
helper processes can be ran after a build given not all DLLs from
vcpkg libs get implicitly copied to the bin folder. (2) Allow
fully background and/or GUI processes to use the Windows Subsystem.
This prevents unnecessarily launching a console for the process, as
we either require no user interaction or the user interaction is
all handled in the GUI.