LibCore: Use WinSock2 closesocket() in PosixSocketHelper::close()
Our Core::System::is_socket() helper is not very robust, as it was
it was incorrectly marking m_fd as a non-socket fd. This meant
we were trying to call CLoseHandle() on a socket which fails.
We can just call closesocket() directly given we are in an
explicit socket-only context.