LibRequests: Decode RequestTimingInfo IPC fields as i64's
This was the root cause of the instability when loading
webpages via HTTP(S) on Windows, where we crashed ~70-80%
of the time. We were using i64's (8 bytes) to encode the fields,
but we were using long's (8 bytes on 64-bit Unix, 4 bytes on Windows)
to decode them. So on Windows we were corrupted every field that
got decoded as we were 4 bytes off every time.