LibGfx: Add Direct3DContext on Windows
For WebGL support on Windows, we need to use angle's D3D11 backend.
We can create a shared D3D11 texture, and open a compatible D3D12
handle to that shared resource. This allows us to use skia's D3D12
backend for our painting surface that displays the content angle
is rendering into the D3D11 texture. So our Direct3DContext handles
both D11 and D12 to support all the scenarios we'll need with a single
context instance.