LibWeb/WebGPU: Implement canvas drawing buffer creation in Dawn
In WebGPU, we supply a texture that render passes draw into. In order
to integrate with the canvas painting surface, we share the texture
memory between the render texture and the painting surface in platform
specific ways. The NativeDrawingBuffer abstraction allows us to hide
the setup of said texture and painting surface for each platform and
each WebGPU backend from the GPUCanvasContext implementation.